MAJOR_NUM=1 MINOR_NUM=6 AUX_NUM=0 force all data type to  use 32bits datatype.

+++ src/algorithm/lookuptable.go
@@ -8,7 +8,7 @@ type LookupTable_t struct {
 	DBKey string
 }
 
-func (o* LookupTable_t)Compute(currentTemperature float64) float64 {
+func (o* LookupTable_t)Compute(currentTemperature float32) float32 {
 	val := currentTemperature
 	if val < 25 {
 		return 20
